src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Conference WMS Backend."
authors = ["gokudaisensei <87324237+gokudaisensei@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.99.1"}
pydantic = "1.10.11"
sqlalchemy = "^2.0.17"
pymysql = "^1.1.0"
tenacity = "^8.2.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
